+++ src/features/XIT/CONTC/ConditionText.vue
@@ -60,6 +60,7 @@ defineProps({
   <template v-else-if="condition.type === 'PRODUCTION_RUN'">Run Production</template>
   <template v-else-if="condition.type === 'START_FLIGHT'">Start Flight</template>
   <template v-else-if="condition.type === 'POWER'">Become Governor</template>
+  <template v-else-if="condition.type === 'REPAIR_SHIP'">Repair Ship</template>
   <template v-else>
     {{ condition.type }}
   </template>